requirements.txt
fastapi==0.115.12
groq==0.24.0
motor==3.7.0
openai==1.78.1
Pillow==11.2.1
pydantic==2.11.4
pymongo==4.12.1
python-dotenv==1.1.0
together==1.5.8
uvicorn==0.34.2
ffmpeg-python==0.2.0